requirements.txt
flask
pillow
torchvision
